Rudimentary Eval
Started on eval function. For now it is just the raw material value. I am going to add piece activity as the main determiner, which consists of 4 (maybe 5) parts: Mobility, Centralization, Attack Potential, Pawn Attacks, and as a stretch goal Piece Coordination.

Also created a new library to implement ray tracing for the pieces, as this will be necessary for the attack potential and pawn attack components.